chore: 프로젝트 초기 설정
